+++ mytrim_ODS.cc
@@ -72,9 +72,9 @@ int main(int argc, char *argv[])
   //trimBase *trim = new trimPrimaries( sample );
 
 
-  //float r = 10.0;
-  float r = atof( argv[2] ); //10.0;
-  float Cbf = atof( argv[3] );
+  //double r = 10.0;
+  double r = atof( argv[2] ); //10.0;
+  double Cbf = atof( argv[3] );
 
   sample->bc[0] = sampleBase::INF; // no PBC in x (just clusterless matrix)
   sample->initSpatialhash( int( sample->w[0] / r ) - 1,
@@ -82,9 +82,9 @@ int main(int argc, char *argv[])
                            int( sample->w[2] / r ) - 1 );
 
 
-  // float atp = 0.1; // 10at% Mo 90at%Cu
-  float v_sam = sample->w[0] * sample->w[1] * sample->w[2];
-  float v_cl = 4.0/3.0 * M_PI * cub(r); 
+  // double atp = 0.1; // 10at% Mo 90at%Cu
+  double v_sam = sample->w[0] * sample->w[1] * sample->w[2];
+  double v_cl = 4.0/3.0 * M_PI * cub(r); 
   int n_cl; // = atp * scoef[29-1].atrho * v_sam / ( v_cl * ( ( 1.0 - atp) * scoef[42-1].atrho + atp * scoef[29-1].atrho ) );
 
   n_cl = v_sam * 1.5e-7 * Cbf ; // Allen08 1.5e-4/nm^3
@@ -189,15 +189,15 @@ int main(int argc, char *argv[])
   // create a FIFO for recoils
   queue<ionBase*> recoils;
 
-  float norm;
-  float jmp = 2.7; // diffusion jump distance
+  double norm;
+  double jmp = 2.7; // diffusion jump distance
   int jumps;
-  float dif[3], dif2[3];
+  double dif[3], dif2[3];
 
   massInverter *m = new massInverter;
   energyInverter *e = new energyInverter;
 
-  float A1, A2, Etot, E1, E2;
+  double A1, A2, Etot, E1, E2;
   int Z1, Z2;
 
   snprintf( fname, 199, "%s.Erec", argv[1] );
@@ -206,14 +206,14 @@ int main(int argc, char *argv[])
   snprintf( fname, 199, "%s.dist", argv[1] );
   FILE *rdist = fopen( fname, "wt" );
 
-  float pos1[3], pos2[3];
+  double pos1[3], pos2[3];
 
   ionBase *ff1, *pka;
   int id = 1;
 
-  float A = 84.0, E = 1.8e6; int Z = 36; // 1.8MeV Kr
-  //float A = 58.0, E = 5.0e6; int Z = 28; // 5MeV Ni
-  //float A = 56.0, E = 5.0e6; int Z = 26; // 5MeV Fe
+  double A = 84.0, E = 1.8e6; int Z = 36; // 1.8MeV Kr
+  //double A = 58.0, E = 5.0e6; int Z = 28; // 5MeV Ni
+  //double A = 56.0, E = 5.0e6; int Z = 26; // 5MeV Fe
 
   // 1000 ions
   for( int n = 0; n < nstep; n++ )
