+++ simconf.h
@@ -4,17 +4,17 @@
 // ZBL coefficients a,b,c,d for all element pairs from Z=1..92
 struct scoefLine {
   char sym[3], name[30];
-  float mm1, m1, mnat, rho, atrho, vfermi, heat, lfctr;
+  double mm1, m1, mnat, rho, atrho, vfermi, heat, lfctr;
 };
 
 struct simconfType {
-  float ed, alfa, alpha, tmin, tau, da, cw;
+  double ed, alfa, alpha, tmin, tau, da, cw;
   int id;
 
   // tables from files
   scoefLine scoef[92];
-  float pcoef[92][8];
-  float snuc[92][92][4];
+  double pcoef[92][8];
+  double snuc[92][92][4];
 
   bool fullTraj;
 
@@ -22,7 +22,7 @@ struct simconfType {
   int vacancies_created;
   double KP_vacancies;
 
-  simconfType( float _alfa = 0.0 );
+  simconfType( double _alfa = 0.0 );
 private:
   void read_scoef();
   void read_snuc();
